Move dataset information into timetable
This allows us to remove a bunch of conditionals regarding whether a DAG
is backed by a timetable or dataset condition, and a weird edge case in
serialization where we don't actually deserialize datasets in a
timetable.

Now datasets are always serialized as a part of the timetable, and we
always evaluate the timetable and datasets. Timetables that do not
actually contain datasets (most of them) simply always evaluate to
False.