Remove unused replace boolean parameter in insert_rows function of OracleHook class
The replace boolean parameter claims to replace rows instead of insert, but this functionality has not been implemented and the documentation claims it works. This can be misleading to any user of this class. Furthermore, the parameter isn't even being used anywhere in this function.

In this commit, I removed the parameter to prevent any confusion amongst the users of this class.